+++ app/badge/page.mdx
@@ -1,10 +1,13 @@
-import {DemoBasicBadge, DemoStrongBadge,DemoSmallBadge} from "./Demo"
+import * as BadgeDemo from "./Demo"
 
 ## Basic 
-<DemoBasicBadge/>
+<BadgeDemo.DemoBasicBadge/>
 
 ## Strong 
-<DemoStrongBadge/>
+<BadgeDemo.DemoStrongBadge/>
 
 ## Small
-<DemoSmallBadge/>
+<BadgeDemo.DemoSmallBadge/>
+
+## Link
+<BadgeDemo.DemoBadgeWithLink/>
